requirements.txt
asgiref==3.2.7
Django==3.0.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
